Refactoring test input creation
