+++ backend/src/main/java/ai/giskard/web/dto/TestInputDTO.java
@@ -1,5 +1,6 @@
 package ai.giskard.web.dto;
 
+import com.dataiku.j2ts.annotations.UINullable;
 import com.fasterxml.jackson.annotation.JsonAlias;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
@@ -11,6 +12,7 @@ public class TestInputDTO {
 
     @NotBlank
     private String name;
+    @UINullable
     private String value;
     private String type;
     @JsonAlias("is_alias")
