+++ frontend/src/generated-sources/ai/giskard/web/dto/ml/dataset-dto.ts
@@ -1,4 +1,4 @@
-import type {ColumnType} from '../../../domain/column-type';
+import type {ColumnType} from './../../../domain/column-type';
 import type {ProjectDTO} from './project-dto';
 
 /**
@@ -6,9 +6,9 @@ import type {ProjectDTO} from './project-dto';
  */
 export interface DatasetDTO {
     columnDtypes: {[key: string]: string};
+    columnTypes: {[key: string]: ColumnType};
     compressedSizeBytes: number;
     createdDate: any /* TODO: Missing translation of java.time.Instant */;
-    columnTypes: {[key: string]: ColumnType};
     id: string;
     name: string;
     originalSizeBytes: number;
