+++ frontend/src/generated-sources/ai/giskard/web/dto/test-input-dto.ts
@@ -5,5 +5,5 @@ export interface TestInputDTO {
     isAlias: boolean;
     name: string;
     type: string;
-    value: string;
+    value?: string | null;
 }
