+++ frontend/src/generated-sources/j2ts-generated-metadata.json
@@ -10,6 +10,7 @@
     "ai.giskard.web.dto.user.PasswordChangeDTO": "PasswordChangeDTO",
     "ai.giskard.web.dto.FeedbackReplyDTO": "FeedbackReplyDTO",
     "ai.giskard.web.dto.ml.write.FilePostDTO": "FilePostDTO",
+    "ai.giskard.domain.ColumnType": "ColumnType",
     "ai.giskard.web.dto.ml.ExecuteTestSuiteRequest": "ExecuteTestSuiteRequest",
     "ai.giskard.web.dto.ml.SuiteTestExecutionDTO": "SuiteTestExecutionDTO",
     "ai.giskard.web.rest.vm.TokenAndPasswordVM": "TokenAndPasswordVM",
@@ -32,7 +33,6 @@
     "ai.giskard.web.dto.ml.CodeBasedTestPresetDTO": "CodeBasedTestPresetDTO",
     "ai.giskard.web.dto.TestSuiteCompleteDTO": "TestSuiteCompleteDTO",
     "ai.giskard.web.dto.SuiteTestDTO": "SuiteTestDTO",
-    "ai.giskard.domain.ColumnType": "ColumnType",
     "ai.giskard.jobs.JobState": "JobState",
     "ai.giskard.web.dto.FeatureMetadataDTO": "FeatureMetadataDTO",
     "ai.giskard.web.dto.GenerateTestSuiteDTO": "GenerateTestSuiteDTO",
